Fix race condition in re/starting API
The API server can take a bit of time to come up after starting or
restarting.  This can cause other handlers to fail with
ConnectionRefused.

When the handler fails within the same hook context as the installation,
it can leave the charm in a state where the next hook will re-attempt
the installation, which isn't handled well.  We need to improve support
for install being called again.